Add a GitHub Action to push releases to PyPI
Create a new `release` workflow based on the [poetry-publish](https://github.com/marketplace/actions/publish-python-poetry-package) action that only runs of all other required workflows complete successfully on a release tag.

Also, upgrade the Python version on the `pre-commit` action to 3.11 for consistency, and ensure that formatting in all workflow files are consistent.