Detect plugins that do not use config rendering (#130)
Distinguish between plugins that do not support configuration rendering, and those that do, by raising NotImplementedError in the former case. Make sure both cases are properly handled.

Post-initialisation rendering already had this change made in the previous PR, but did not have the correct error handling. This PR addresses that.