Add built-in support for handling `NameEmail` attribute types
Decode, encode and format `NameEmail` type attributes correctly for most applications.

This fixes uses of the attribute type in plugins that do not provide explicit encoding/decoding for this type.

This will not be done for every possible type, but `NameEmail` is common and generic enough that it makes the cut.